lil styling and refactor
